Add files via upload
About the problem
It is undeniable that obtaining the proper mobile health application is becoming more difficult with the passage of time. This complication poses numerous issues for various stakeholders, making it difficult to identify appropriate technology/apps to assist them in enhancing their health application usage. This project will attempt to answer the question “How effective are mobile health apps in regard to different usage and what impact does different geographical locations have on the efficiency of these applications.”